New kernel flasher and wallpaper picker
Added a modified AnyKernel3 as alternative kernel flasher for devices with A/B partition scheme.
To select the alternative kernel flasher, add "flasher = anykernel" to devices.cfg.
When using "anykernel" it is required to use a comma separated list of devicen names

Added option to set the screen resolution of the device in devices.cfg for devices that do not support the automatic detection of the resolution or to overwrite the resolution detected.
To set the resolution, add "resolution = <width>x<height>" to devices.cfg

Signed-off-by: Re4son <3520622-re4son@users.noreply.gitlab.com>